# PERF: Block.apply